requirements.txt
Flask==3.0.3
python-chess==1.999